Testing, bugfixes, and v0.0.1b1 (#5)
* add park.cfg

* we don't need the AWS passenv values

* move BootstrapInstall into pipipxx.py

* disable in virtualenv

* add integration testing bash script and Travis handlers

* add more verbose error messaging to setup

* add path check

* declare bankruptcy, move code into setup.py, and give up on unit tests

* fix test file path

* remove now-irrelevant test helper

* userpath verify is both unnecessary and introducing errors. userpath add already verifies, then adds if verification fails

* add fixes to changelog

* remove hard Python 3.6 requirement: delegating that to pipx

* bump version to 0.0.1b1
