+++ .travis/install-check.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+tox -e build && ./test/test_pipipxx.sh "$(pwd)/dist"
