--- examples/__init__.py
@@ -1,1 +1,0 @@
-"""Stub to allow relative imports of examples from tests."""
