--- examples/src/__init__.py
@@ -1,1 +1,0 @@
-"""Write some examples and put them in this directory."""
