--- examples/test/__init__.py
@@ -1,1 +1,0 @@
-"""Put tests for the examples in this directory."""
