--- examples/test/test_make_tests.py
@@ -1,8 +1,0 @@
-"""Placeholder module to remind you to write tests."""
-import pytest
-
-
-@pytest.mark.xfail(strict=True)
-@pytest.mark.examples
-def test_write_tests():
-    assert False
