+++ park.cfg
@@ -1,0 +1,18 @@
+[DEFAULT]
+version: 0.0.1
+author: Matt Bullock
+author_email: m@ttsb42.com
+url: https://pipipxx.readthedocs.io/en/stable/
+description: Did you mean to install pipipxx?
+long_description:
+    This package has been parked by {author} to protect you against packages
+    adopting names that might be common mistakes when looking for ours. You probably
+    wanted to install pipipxx. For more information, see {url}.
+description_keys:
+    author
+    url
+
+[names]
+pipxinpipx:
+pipx-in-pipx:
+pipipx:
