+++ setup.cfg
@@ -47,4 +47,4 @@ force_grid_wrap = 0
 combine_as_imports = True
 not_skip = __init__.py
 known_first_party = pipipxx
-known_third_party =pytest,setuptools
+known_third_party =setuptools
