--- test/__init__.py
@@ -1,1 +1,0 @@
-"""Stub to allow relative imports between test groups."""
