--- test/requirements.txt
@@ -1,4 +1,0 @@
-mock
-pytest>=3.3.1
-pytest-cov
-pytest-mock
