--- test/vectors/README.rst
@@ -1,5 +1,0 @@
-*******************
-Static Test Vectors
-*******************
-
-Any static test vectors should be placed in this directory.
