+++ tox.ini
@@ -1,9 +1,7 @@
 [tox]
 envlist =
-    autoformat,
-    py{36,37}-{local,integ,examples},
-    noenvvars, sourcebuildcheck,
-    {flake8,pylint}{,-tests,-examples},
+    lint,
+    {flake8,pylint},
     mypy,
     bandit,
     doc8, readme, docs,
@@ -32,51 +30,11 @@ skipsdist = True
 # Tests #
 #########
 
-[testenv:base-command]
-commands = pytest --basetemp={envtmpdir} -l --cov pipipxx {posargs}
-
 [testenv]
 skip_install = True
-passenv =
-    # Pass through AWS credentials
-    AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY AWS_SESSION_TOKEN \
-    # Pass through AWS profile name (useful for local testing)
-    AWS_PROFILE \
-    # Pass through the default AWS region (used for integration tests)
-    AWS_DEFAULT_REGION
 sitepackages = False
-deps = -rtest/requirements.txt
-commands =
-    # Local tests: no network access required
-    local: {[testenv:base-command]commands} test/ -m local
-    # Integration tests: requires network access and might require service credentials
-    integ: {[testenv:base-command]commands} test/ -m integ
-    # Acceptance tests: testing against static test vectors : same requirements as integ
-    accept: {[testenv:base-command]commands} test/ -m accept
-    # Test the examples : same requirements as integ
-    examples: {[testenv:base-command]commands} examples/test/ -m examples
-    # Run all known tests : same requirements as integ
-    all: {[testenv:base-command]commands} test/ examples/test/
-    # You decide what tests to run
-    manual: {[testenv:base-command]commands}
 
-# Verify that local tests work without environment variables present
-[testenv:noenvvars]
-basepython = python3
-sitepackages = False
-deps = {[testenv]deps}
-commands = {[testenv:base-command]commands} test/ -m local
-
-# Verify that tests can be successfully run from the source build.
-[testenv:sourcebuildcheck]
-basepython = python3
-sitepackages = False
-recreate = True
-deps =
-    {[testenv:build]deps}
-commands =
-    {[testenv:build]commands}
-    {toxinidir}/test/source-build-check.sh {envtmpdir} {toxinidir}/dist
+# pipipxx tests cannot run inside virtual environments :(
 
 ###############
 # Type checks #
@@ -103,7 +61,7 @@ deps =
 commands =
     python -m mypy \
         --linecoverage-report build \
-        src/pipipxx.py \
+        setup.py \
         {posargs}
     {[testenv:mypy-coverage]commands}
 
@@ -121,65 +79,19 @@ deps =
     flake8-bugbear
 commands =
     flake8 \
-        src/pipipxx.py \
         setup.py \
         doc/conf.py
-
-[testenv:flake8-tests]
-basepython = {[testenv:flake8]basepython}
-deps = {[testenv:flake8]deps}
-commands =
-    flake8 \
-        # Ignore F811 redefinition errors in tests (breaks with pytest-mock use)
-        # Ignore F841 local variable assigned but never used (useful for collecting locals for test reports)
-        # Ignore D101,D102,D103 docstring requirements for tests
-        # Ignore B011 assert False warning (never run tests with python -O)
-        --ignore F811,F841,D101,D102,D103,B011 \
-        test/
-
-[testenv:flake8-examples]
-basepython = {[testenv:flake8]basepython}
-deps = {[testenv:flake8]deps}
-commands =
-    flake8 \
-        examples/src/
-    flake8 \
-        # Ignore F811 redefinition errors in tests (breaks with pytest-mock use)
-        # Ignore F841 local variable assigned but never used (useful for collecting locals for test reports)
-        # Ignore D101,D102,D103 docstring requirements for tests
-        # Ignore B011 assert False warning (never run tests with python -O)
-        --ignore F811,F841,D101,D102,D103,B011 \
-        examples/test/
 
 [testenv:pylint]
 basepython = python3
 deps =
-    {[testenv]deps}
     pyflakes
     pylint>=2.0.0
 commands =
     pylint \
-        --rcfile=src/pylintrc \
-        src/pipipxx.py  \
+        --rcfile=pylintrc \
         setup.py \
         doc/conf.py
-
-[testenv:pylint-tests]
-basepython = {[testenv:pylint]basepython}
-deps = {[testenv:pylint]deps}
-commands =
-    pylint \
-        --rcfile=test/pylintrc \
-        test/unit/ \
-        test/functional/ \
-        test/integration/
-
-[testenv:pylint-examples]
-basepython = {[testenv:pylint]basepython}
-deps = {[testenv:pylint]deps}
-commands =
-    pylint --rcfile=examples/src/pylintrc examples/src/
-    pylint --rcfile=examples/test/pylintrc examples/test/
 
 [testenv:bandit]
 basepython = python3
@@ -187,13 +99,13 @@ deps = bandit
 commands = bandit \
     # B322: Ignore Python 2 input check: we only support Python 3
     -s B322 \
-    -r src/pipipxx.py
+    -r setup.py
 
 # Prone to false positives: only run manually
 [testenv:vulture]
 basepython = python3
 deps = vulture
-commands = vulture src/pipipxx.py
+commands = vulture setup.py
 
 [testenv:blacken-src]
 basepython = python3
@@ -201,11 +113,8 @@ deps =
     black
 commands =
     black --line-length 120 \
-        src/pipipxx.py \
         setup.py \
         doc/conf.py \
-        test/ \
-        examples/ \
         {posargs}
 
 [testenv:blacken]
@@ -225,9 +134,7 @@ basepython = python3
 deps = isort
 commands = isort \
     -rc \
-        src \
         test \
-        examples \
         doc \
         setup.py \
     {posargs}
@@ -264,22 +171,14 @@ whitelist_externals = {[testenv:resetdocs]whitelist_externals}
 deps =
     {[testenv:autoformat]deps}
     {[testenv:flake8]deps}
-    {[testenv:flake8-tests]deps}
-    {[testenv:flake8-examples]deps}
     {[testenv:pylint]deps}
-    {[testenv:pylint-tests]deps}
-    {[testenv:pylint-examples]deps}
     {[testenv:doc8]deps}
     {[testenv:readme]deps}
     {[testenv:bandit]deps}
 commands =
     {[testenv:autoformat]commands}
     {[testenv:flake8]commands}
-    {[testenv:flake8-tests]commands}
-    {[testenv:flake8-examples]commands}
     {[testenv:pylint]commands}
-    {[testenv:pylint-tests]commands}
-    {[testenv:pylint-examples]commands}
     {[testenv:doc8]commands}
     {[testenv:readme]commands}
     {[testenv:bandit]commands}
@@ -336,7 +235,6 @@ commands =
 [testenv:docs]
 basepython = python3
 deps =
-    {[testenv]deps}
     -r{toxinidir}/doc/requirements.txt
 commands =
     {[testenv:docs-build]commands}
